+++ src/common.c
