+++ src/snapshot.c
