+++ src/util.c
